Make sure other methods don't get caught in our checks
The method call check was too loose and would check on all method calls
on your object. For Tactician itself, this was fine since there were no
other methods and it was a final class. However, it was a problem on
custom command buses where you could have other methods OR use a mock
object which has other methods to set expectations. In that case, you'd
get things like trying to looking for a PHPUnit
"NumberOfTimesInvokedHandler but could not be found for
NumberOfTimesInvoked command" because it was trying to take self::once()
as a command on the method $mock->expects