Skip trying to resolve interfaces or abstract classes
They won't be the concrete classes we try to resolve on, so there's no
point in trying to check them anyways. This most often happens when the
Command used a typehint for some utility like a queue and we've lost the
ability to typecheck there anyways without some very complicated
trickery.