Check that class exists in broker before getting it
In the InterfacesAndAbstractClassesShouldNotBeReported testcase, trying
to fetch an abstract or interface class via the broker's getClass()
method raises a PHPStan error. To prevent this, we now check if the
class exists in the broker, via a hasClass() call. I've kept the
additional isInterface/isAbstract calls, however, as I'm not completely
sure about the intended behavior on PHPStan's side and need to research
further.